+++ src/main/java/com/skycat/mystical/common/spell/consequence/DisableDaylightBurningConsequence.java
@@ -11,11 +11,11 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.Random;
 
-public class DisableDaylightBurningConsequence extends SpellConsequence { // TODO: Config
+public class DisableDaylightBurningConsequence extends SpellConsequence {
     public static Factory FACTORY = new Factory();
 
     protected DisableDaylightBurningConsequence() {
-        super(DisableDaylightBurningConsequence.class, null, 15d); // TODO: Scaling
+        super(DisableDaylightBurningConsequence.class, null, 15d);
     }
 
     @Override
@@ -53,7 +53,7 @@ public class DisableDaylightBurningConsequence extends SpellConsequence { // TOD
         }
 
         @GameTest(templateName = TestUtils.BORDERED_BARRIER_BOX)
-        public void testHaven(TestContext context) { // TODO: Test
+        public void testHaven(TestContext context) {
             setUpTest(context);
             TestUtils.havenAll(context);
             context.waitAndRun(75, () -> {
@@ -63,7 +63,7 @@ public class DisableDaylightBurningConsequence extends SpellConsequence { // TOD
         }
 
         @GameTest(templateName = TestUtils.BORDERED_BARRIER_BOX)
-        public void testHavenAndSpell(TestContext context) { // TODO: Test
+        public void testHavenAndSpell(TestContext context) {
             setUpTest(context);
             TestUtils.havenAll(context);
             Mystical.getSpellHandler().activateNewSpellWithConsequence(this);