+++ src/main/java/com/skycat/mystical/common/spell/consequence/LevitateConsequence.java
@@ -53,7 +53,7 @@ public class LevitateConsequence extends SpellConsequence implements EntitySleep
     }
 
     public LevitateConsequence(int length, int level, Class callbackType) {
-        super(LevitateConsequence.class, callbackType, 30d); // TODO: Scaling
+        super(LevitateConsequence.class, callbackType, 30d);
         this.length = length;
         this.level = level;
     }