+++ src/main/java/com/skycat/mystical/common/spell/consequence/SheepColorChangeConsequence.java
@@ -7,7 +7,7 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.Random;
 
-public class SheepColorChangeConsequence extends SpellConsequence { // TODO: Tests // TODO: Config
+public class SheepColorChangeConsequence extends SpellConsequence { // TODO: Tests
     public static final Factory FACTORY = new Factory();
 
     @Override