+++ src/main/java/com/skycat/mystical/common/spell/consequence/TurboMobsConsequence.java
@@ -13,7 +13,7 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.Random;
 
-public class TurboMobsConsequence extends SpellConsequence { // TODO: Maybe make this work with wardens?  // TODO: Tests
+public class TurboMobsConsequence extends SpellConsequence { // TODO: Tests
     public static final Factory FACTORY = new Factory();
     @Getter public EntityType<?> entityType;
 
@@ -44,7 +44,7 @@ public class TurboMobsConsequence extends SpellConsequence { // TODO: Maybe make
             if (entry.isPresent()) {
                 EntityType<?> type = entry.get().value();
                 SpawnGroup spawnGroup = type.getSpawnGroup();
-                if (spawnGroup != null && spawnGroup != SpawnGroup.MISC && type != EntityType.GIANT) { // TODO: Use EntityTypePredicate, TODO: config
+                if (spawnGroup != null && spawnGroup != SpawnGroup.MISC && type != EntityType.GIANT) {
                     return new TurboMobsConsequence(type);
                 } else {
                     Utils.log("Making TurboMobsConsequence: skipping unspawnable type: " + type.getName().getString(), LogLevel.INFO);